Move color_cycler, game_clock, memory_tiered_tasks
Move color_cycler, game_clock, and memory_tiered_tasks to mapping_core
