+++ src/chkdraft/mapping/sc_gl_graphics.cpp
@@ -1,11 +1,11 @@
 #include "sc_gl_graphics.h"
 #include <chkdraft/ui/main_windows/gui_map.h>
+#include <mapping_core/render/memory_tiered_tasks.h>
 #include "cross_cut/logger.h"
 #include <rarecpp/json.h>
 #include <chrono>
 #include <cstring>
 #include <Windows.h>
-#include "memory_tiered_tasks.h"
 #include <glad/glad.h>
 #include "gl/base.h"
 #include <glm/vec3.hpp>
