+++ src/mapping_core/render/color_cycler.h
